Migrate regressions from 4501-5000
