+++ spacy/tests/test_cli.py
@@ -1,28 +1,67 @@
-import pytest
-from click import NoSuchOption
-from packaging.specifiers import SpecifierSet
-from spacy.training import docs_to_json, offsets_to_biluo_tags
-from spacy.training.converters import iob_to_docs, conll_ner_to_docs, conllu_to_docs
-from spacy.schemas import ProjectConfigSchema, RecommendationSchema, validate
-from spacy.lang.nl import Dutch
-from spacy.util import ENV_VARS, load_model_from_config
-from spacy.cli import info
-from spacy.cli.init_config import init_config, RECOMMENDATIONS
-from spacy.cli._util import validate_project_commands, parse_config_overrides
-from spacy.cli._util import load_project_config, substitute_project_variables
-from spacy.cli._util import is_subpath_of
-from spacy.cli._util import string_to_list
-from spacy import about
-from spacy.util import get_minor_version
-from spacy.cli.validate import get_model_pkgs
-from spacy.cli.download import get_compatibility, get_version
-from spacy.cli.package import get_third_party_dependencies
-from thinc.api import ConfigValidationError, Config
-import srsly
 import os
 
-from .util import make_tempdir
+import pytest
+import srsly
+from click import NoSuchOption
+from packaging.specifiers import SpecifierSet
+from thinc.api import Config, ConfigValidationError
+
+from spacy import about
+from spacy.cli import info
+from spacy.cli._util import is_subpath_of, load_project_config
+from spacy.cli._util import parse_config_overrides, string_to_list
+from spacy.cli._util import substitute_project_variables
+from spacy.cli._util import validate_project_commands
+from spacy.cli.download import get_compatibility, get_version
+from spacy.cli.init_config import RECOMMENDATIONS, init_config
+from spacy.cli.package import get_third_party_dependencies
+from spacy.cli.validate import get_model_pkgs
+from spacy.lang.nl import Dutch
+from spacy.language import Language
+from spacy.schemas import ProjectConfigSchema, RecommendationSchema, validate
+from spacy.training import Example, docs_to_json, offsets_to_biluo_tags
+from spacy.training.converters import conll_ner_to_docs, conllu_to_docs
+from spacy.training.converters import iob_to_docs
+from spacy.util import ENV_VARS, get_minor_version, load_model_from_config
+
 from ..cli.init_pipeline import _init_labels
+from .util import make_tempdir
+
+
+@pytest.mark.issue(4665)
+def test_issue4665():
+    """
+    conllu_to_docs should not raise an exception if the HEAD column contains an
+    underscore
+    """
+    input_data = """
+1	[	_	PUNCT	-LRB-	_	_	punct	_	_
+2	This	_	DET	DT	_	_	det	_	_
+3	killing	_	NOUN	NN	_	_	nsubj	_	_
+4	of	_	ADP	IN	_	_	case	_	_
+5	a	_	DET	DT	_	_	det	_	_
+6	respected	_	ADJ	JJ	_	_	amod	_	_
+7	cleric	_	NOUN	NN	_	_	nmod	_	_
+8	will	_	AUX	MD	_	_	aux	_	_
+9	be	_	AUX	VB	_	_	aux	_	_
+10	causing	_	VERB	VBG	_	_	root	_	_
+11	us	_	PRON	PRP	_	_	iobj	_	_
+12	trouble	_	NOUN	NN	_	_	dobj	_	_
+13	for	_	ADP	IN	_	_	case	_	_
+14	years	_	NOUN	NNS	_	_	nmod	_	_
+15	to	_	PART	TO	_	_	mark	_	_
+16	come	_	VERB	VB	_	_	acl	_	_
+17	.	_	PUNCT	.	_	_	punct	_	_
+18	]	_	PUNCT	-RRB-	_	_	punct	_	_
+"""
+    conllu_to_docs(input_data)
+
+
+@pytest.mark.issue(4924)
+def test_issue4924():
+    nlp = Language()
+    example = Example.from_dict(nlp.make_doc(""), {})
+    nlp.evaluate([example])
 
 
 def test_cli_info():
